Ensure that navigation bar setting survives a reboot
The problem -- If you turn off the navbar and then reboot, the
navbar would get recreated on boot and so show a false state for
the navbar switch.

Although not used by majority of users, a few users like to turn
off the navbar so they can use things like Pie, OP gestures, etc.

Change-Id: Idb8b7088737c4896775ec5405bdd9238b0fa475e