Navbar: Make sure navbar updates correctly
On devices that don't have a navbar by default (OP3/3T, OP5, etc) we
come across some jank. On boot, you can toggle the navbar on and off
as much as you want BUT once you change the density and go to toggle
the navbar on and off, the navbar starts to flicker a bit. You see it
pop up for a second but all black and then it disappears.

Changing the order in which things are called, fixes this.

Change-Id: Iaaa46cc773d4553cf4f4f3bb919cf82b59435830